renamed NewStepper to Stepper component
