+++ src/routes/CreateSafePage/steps/OwnersAndConfirmationsNewSafeStep.tsx
@@ -14,7 +14,7 @@ import Button from 'src/components/layout/Button'
 import Field from 'src/components/forms/Field'
 import ButtonHelper from 'src/components/ButtonHelper'
 import SelectField from 'src/components/forms/SelectField'
-import { useStepper } from 'src/components/NewStepper/stepperContext'
+import { useStepper } from 'src/components/Stepper/stepperContext'
 import { providerNameSelector } from 'src/logic/wallets/store/selectors'
 import { disabled, extraSmallFontSize, lg, sm, xs } from 'src/theme/variables'
 import Hairline from 'src/components/layout/Hairline'
